2-Python-Terraform/requirements.txt
python-dotenv==1.0.1
python-terraform==0.10.1
